Feature: Pseudo partial reading into ismrmrd files (MR raw data)
Every bit of data is read and loaded because I did not find a way to only read the hard-header part using h5. However, in order to save RAM, instead of allocating and filling the full kspace volume, only what's necessary is allocated. This is very useful when only part of the dataset is needed (subpart of the kspace, specific contrast or channel...). There's also a mode for loading the data in a compact way (i.e., without zero-filling missing lines), with parallel cartesian sampling in mind.